backend/requirements.txt
haystack-ai==2.9.0
google-generativeai==0.8.3
flask==3.0.3
flask-cors==5.0.0
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.10
pgvector==0.3.6
sqlalchemy==2.0.36
pydantic==2.10.3
python-dotenv==1.0.1
reportlab==4.2.5
pytest==8.3.4
pytest-flask==1.3.0
sentence-transformers==3.3.1
